Display fatal error if no available sms client
