+++ resources/authgear/templates/en/translation.json
@@ -1017,6 +1017,7 @@
   "v2.error.verification-code-invalid": "Verification code is incorrect.",
   "v2.error.verification-code-required": "Please enter the verification code you just received.",
   "v2.error.web-ui-invalid-session-return": "You may have pressed the back button, or there is some issue with cookies. Try: <ol><li>Performing the action again from the application.</li><li>Retrying after clearing cookies.</li><li> If the problem persists, please contact the administrator.</li></ol>",
+  "v2.error.no-available-sms-client": "There could be a misconfiguration in the system. Please contact the systems admin.",
   "v2.error.webhook-delivery-timeout": "Operation is disallowed because the webhook delivery resulted in timeout.",
   "v2.error.webhook-disallowed": "Operation is disallowed",
   "v2.error.webhook-invalid-response": "[To Developer] Your webhook handler did NOT return a valid response. Please visit <a class=\"link\" target=\"_blank\" href=\"https://docs.authgear.com/webhooks/webhooks\">https://docs.authgear.com/webhooks/webhooks</a>",
@@ -1078,6 +1079,7 @@
   "v2.page.enter-totp.reauth.description": "For your security, we need to verify it''s you. Enter the 6-digit code you see in the authenticator device/app.",
   "v2.page.fatal-error.default.title": "Unexpected Issues",
   "v2.page.fatal-error.token-invalid.title": "Page expired or invalid",
+  "v2.page.fatal-error.no-available-sms-client.title": "Fail to send SMS",
   "v2.page.forgot-password-link-sent.default.description": "A link to reset your password has been sent to {MaskedLoginID}",
   "v2.page.forgot-password-link-sent.default.resend": "Resend",
   "v2.page.forgot-password-link-sent.default.resent-countdown-label-unit": "Resend in %s",
