+++ resources/authgear/templates/en/web/authflowv2/fatal_error.html
@@ -8,6 +8,9 @@
     {{ if eq .Error.reason "AccountManagementTokenInvalid" }}
       {{ $title = include "v2.page.fatal-error.token-invalid.title" nil }}
     {{ end }}
+    {{ if eq .Error.reason "NoAvailableSMSClient" }}
+      {{ $title = include "v2.page.fatal-error.no-available-sms-client.title" nil }}
+    {{ end }}
   {{ end }}
 
   {{ $error_message := include "v2.error.server" nil }}
@@ -18,6 +21,8 @@
       {{ $error_message = include "v2.error.no-public-signup" nil }}
     {{ else if eq .Error.reason "AccountManagementTokenInvalid" }}
       {{ $error_message = include "v2.error.settings-token-invalid" nil }}
+    {{ else if eq .Error.reason "NoAvailableSMSClient" }}
+      {{ $error_message = include "v2.error.no-available-sms-client" nil }}
     {{ end }}
   {{ end }}
   {{ template "authflowv2/__error_page_layout.html"
